% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity1_data_entry.R
\name{activity1_data_entry}
\alias{activity1_data_entry}
\title{Data entry for Activity 1}
\usage{
activity1_data_entry(nthrow = 50)
}
\arguments{
\item{nthrow}{Number of imaginary throws of coin per participant}
}
\value{
array of characters, one row per participant, one column per throw, derived from data entered at console
}
\description{
Simplifies data entry for Activty 1 in Chapter 1 of Playing With Statistics
}
\examples{
activity1_data_entry(5) # enter data when each participant has made just 5 imaginary throws of coin

}
