% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity1_data_sm.R
\docType{data}
\name{activity1_data_sm}
\alias{activity1_data_sm}
\title{Results from Smartodds version of Activity 1}
\format{
A data frame with 48 rows and 50 columns
Each row corresponds to one participant
The columns show the sequence of Heads ("H") and Tails ("T") provided by each participant
}
\source{
{in-house}
}
\usage{
activity1_data_sm
}
\description{
Data collected from Smartodds employees carrying out Activity 1 from Playing With Statistics
}
\examples{
apply(activity1_data_sm, 1, function(x) sum(x=="H")) \%>\% unlist # count number of heads generated by each participant


}
\keyword{datasets}
