% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity1_run_app.R
\name{activity1_run_app}
\alias{activity1_run_app}
\title{Run  App 1
Executes app for Activity 1 in Playing With Statistics}
\usage{
activity1_run_app()
}
\description{
Run  App 1
Executes app for Activity 1 in Playing With Statistics
}
\examples{
activity1_run_app()

}
