% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity1_s2_analysis.R
\name{activity1_s2_analysis}
\alias{activity1_s2_analysis}
\title{Analyse Statistic 2 for Activity 1}
\usage{
activity1_s2_analysis(
  data_mat,
  n_throws = 50,
  x_low = 1,
  x_up = 20,
  include_data = TRUE,
  print_summary = TRUE
)
}
\arguments{
\item{data_mat}{Matrix containing data (H/T): one row per individual, one column per imaginary coin toss}

\item{n_throws}{Number of throws per individual}

\item{x_low}{Lower limit for plotting}

\item{x_up}{Upper limit for plotting}

\item{include_data}{include data in plots? (TRUE/FALSE)}

\item{print_summary}{print summary of results?  (TRUE/FALSE)}
}
\value{
list containing theoretical and sample statistics for statistic 2 in activity 1
}
\description{
Analyses Statistic 2 for Activity 1 as described in Chapter 1 of Playing With Statistics
}
\examples{
activity1_s2_analysis(activity1_data_sm)

}
