% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity1_stats.R
\name{activity1_stats}
\alias{activity1_stats}
\title{Statistics for Activity 1}
\usage{
activity1_stats(x)
}
\arguments{
\item{x}{vector sequence of H and/or T}
}
\description{
Calculates Statistics 1 and 2 for Activity 1 from Playing With Statistics
}
\examples{
x <- c('H','H','T','T','T','H')
activity1_stats(x)

}
