% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_game_sim.R
\name{activity2_game_sim}
\alias{activity2_game_sim}
\title{Activity 2 game simulation}
\usage{
activity2_game_sim(
  n_rounds = 30,
  plot = TRUE,
  n_bins = 10,
  Theta = NULL,
  rand_Theta = FALSE,
  show_Theta = FALSE,
  seed = NULL
)
}
\arguments{
\item{n_rounds}{Number of rounds of simulated dice throws}

\item{plot}{Should histogram be drawn (TRUE/FALSE)?}

\item{n_bins}{Number of bins for (optional) histogram}

\item{Theta}{True value of Theta}

\item{rand_Theta}{randomise Theta (TRUE/FALSE)? if TRUE, Theta argument is ignored}

\item{show_Theta}{show Theta in output (TRUE/FALSE)?}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
list containing simulated scores and their mean value
}
\description{
Generates data from which your aim is to guess the parameter Theta, as described in Activity 2 of Playing With Statistics
}
\examples{
activity2_game_sim()

}
