% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_game_sim_rep.R
\name{activity2_game_sim_rep}
\alias{activity2_game_sim_rep}
\title{Activity 2 repeat game simulation}
\usage{
activity2_game_sim_rep(
  n_rep = 1000,
  n_rounds = 30,
  Theta = NULL,
  rand_Theta = FALSE,
  show_estimates = FALSE,
  seed = NULL
)
}
\arguments{
\item{n_rep}{Number of repeats of game}

\item{n_rounds}{Number of rounds ine each game of simulated dice throws}

\item{Theta}{True value of Theta}

\item{rand_Theta}{randomise Theta (TRUE/FALSE)? if TRUE, Theta argument is ignored}

\item{show_estimates}{(TRUE/FALSE)? if TRUE, include estimates as part of output}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
list containing simulated scores and their mean value
}
\description{
Repeats the game described in Activity 2 of Playing With Statistics multiple times
}
\examples{
activity2_game_sim_rep()

}
