% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_resample_est.R
\name{activity2_resample_est}
\alias{activity2_resample_est}
\title{Activity 2 resampling of estimates}
\usage{
activity2_resample_est(dice_scores, nrep = 10000, plot = TRUE, seed = NULL)
}
\arguments{
\item{dice_scores}{Values of original sample of scores}

\item{nrep}{Number of resamples}

\item{plot}{Should boxplot be drawn (TRUE/FALSE)?}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
invisible list containing mean, standard deviation, enumeration and quantiles of estimates based on resamples
}
\description{
Generates boxplot of Theta estimates based on resampling, as described in Activity 2 of Playing With Statistics
}
\examples{

estimates <- activity2_data_sim(plot = FALSE)
activity2_resample_est(estimates$score)
}
