% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_run_app.R
\name{activity2_run_app}
\alias{activity2_run_app}
\title{Run  App 2
Executes app for Activity 2 in Playing With Statistics}
\usage{
activity2_run_app()
}
\description{
Run  App 2
Executes app for Activity 2 in Playing With Statistics
}
\examples{
activity2_run_app()

}
