% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_sd_by_size.R
\name{activity2_sd_by_size}
\alias{activity2_sd_by_size}
\title{Analyse Activity 2 with Different Number of Throws}
\usage{
activity2_sd_by_size(
  n_throws = c(10, 25, 50, 100, 250, 500, 1000, 2500, 5000),
  nrep = 10000,
  Theta = 7.5,
  transform_sd = F,
  log_trans = F,
  seed = NULL
)
}
\arguments{
\item{n_throws}{Vector of numbers of simulated dice throws}

\item{nrep}{Number of repetitions for each experiment}

\item{Theta}{True value of Theta}

\item{transform_sd}{Transform Standard Deviation? (True/False)}

\item{log_trans}{Plot both axes on log scale? (True/False)}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
Boxplots (as defined in Playing With Statistics) showing distribution of estimator in simulated repeats of Activity 2 with different numbers of throws of dice
}
\description{
Shows effect of sample size on standard deviation of estimate in Activity 2 of Playing With Statistics
}
\examples{
activity2_sd_by_size()

}
