% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_se.R
\name{activity2_se}
\alias{activity2_se}
\title{Activity 2 resampling to obtain standard error of estimates}
\usage{
activity2_se(
  dice_scores,
  nrep = 10000,
  alpha = 0.95,
  plot = FALSE,
  width = 0.1,
  seed = NULL,
  print_results = TRUE
)
}
\arguments{
\item{dice_scores}{Values of original sample of scores}

\item{nrep}{Number of resamples}

\item{alpha}{confidence level}

\item{plot}{Should confidence interval be drawn (TRUE/FALSE)?}

\item{width}{width of confidence interval error bars}

\item{seed}{Set seed to enable identical simulation across calls}

\item{print_results}{Print standard error and confidence intervals to screen (TRUE/FALSE)?}
}
\value{
invisible list containing mean, standard deviation, enumeration and quantiles of estimates based on resamples
}
\description{
Generates error plot of Theta estimates based on resampling, as described in Activity 2 of Playing With Statistics
}
\examples{

estimates <- activity2_game_sim(plot = FALSE)
activity2_se(estimates$score)
}
