% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity2_se_by_size.R
\name{activity2_se_by_size}
\alias{activity2_se_by_size}
\title{Analyse Activity 2 with Different Number of Throws}
\usage{
activity2_se_by_size(
  n_throws = c(10, 25, 50, 100, 250, 500, 1000, 2500, 5000),
  nrep = 10000,
  Theta = 7.5,
  transform_se = F,
  log_trans = F,
  seed = NULL
)
}
\arguments{
\item{n_throws}{Vector of numbers of simulated dice throws}

\item{nrep}{Number of repetitions for each experiment}

\item{Theta}{True value of Theta}

\item{log_trans}{Plot both axes on log scale? (True/False)}

\item{seed}{Set seed to enable identical simulation across calls}

\item{transform_sd}{Transform Standard Error? (True/False)}
}
\value{
Standard errors calculated by resampling of estimates of Theta in Activity 2 with different numbers of throws of dice
}
\description{
Shows effect of sample size on standard error of estimate in Activity 2 of Playing With Statistics
}
\examples{
activity2_se_by_size()

}
