% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity3_execute.R
\name{activity3_execute}
\alias{activity3_execute}
\title{Execute Activity 3}
\usage{
activity3_execute(bank = 25, bankmax = 250)
}
\arguments{
\item{bank}{Initial bank size}

\item{bankmax}{Maximum value of bank. If reached execution terminates.}
}
\value{
summary information from execution of Activity 3
}
\description{
Individual execution of Activity 3 of Playing With Statistics
}
\examples{
activity3_execute()

}
