% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity3_run_app.R
\name{activity3_run_app}
\alias{activity3_run_app}
\title{Run  App 3
Executes app for Activity 3 in Playing With Statistics}
\usage{
activity3_run_app()
}
\description{
Run  App 3
Executes app for Activity 3 in Playing With Statistics
}
\examples{
activity3_run_app()

}
