% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity4_execute.R
\name{activity4_execute}
\alias{activity4_execute}
\title{Execute Activity 4}
\usage{
activity4_execute(bank = 25, bankmax = 250)
}
\arguments{
\item{bank}{Initial bank size}

\item{bankmax}{MAximum value of bank. If reached execution terminates.}
}
\value{
summary information from execution of activity 4
}
\description{
Individual execution of Activity 4 of Playing With Statistics
}
\examples{
activity4_execute()

}
