% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity4_responses_sm.R
\docType{data}
\name{activity4_responses_sm}
\alias{activity4_responses_sm}
\title{Quiz Answers for Activity 4}
\format{
A data frame with 260 rows and 4 columns:
\itemize{
\item team: team number
\item question: question number
\item G: best guess
\item S: claimed measure of accuracy
}
}
\source{
{in-house}
}
\usage{
activity4_responses_sm
}
\description{
Answers to Quiz Questions in Activity 4 of Playing With Statistics
}
\examples{
subset(activity4_responses_sm, team == 1) # data from team 1


}
\keyword{datasets}
