% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity4_run_app.R
\name{activity4_run_app}
\alias{activity4_run_app}
\title{Run  App 4
Executes app for Activity 4 in Playing With Statistics}
\usage{
activity4_run_app()
}
\description{
Run  App 4
Executes app for Activity 4 in Playing With Statistics
}
\examples{
activity4_run_app()

}
