% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_ci_by_size.R
\name{activity5_ci_by_size}
\alias{activity5_ci_by_size}
\title{Analyse Activity 5 with Different Number of Throws}
\usage{
activity5_ci_by_size(
  n_rounds = c(10, 25, 50, 100, 250, 500, 1000, 2500, 5000),
  nrep = 10000,
  Theta = 7.5,
  alpha = 0.95,
  width = 0.1,
  epsilon = 3,
  seed = NULL
)
}
\arguments{
\item{n_rounds}{Vector of rounds of simulated dice throws}

\item{nrep}{Number of repetitions for each experiment}

\item{Theta}{True value of Theta}

\item{alpha}{confidence level}

\item{width}{width of error bars}

\item{epsilon}{Determines scale of plot}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
Confidence intervals for Theta in simulated repeats of Activity 2 with different numbers of throws of dice
}
\description{
Shows effect of sample size on distribution of estimate in Activity 5 of Playing With Statistics
}
\examples{
activity5_ci_by_size()

}
