% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_matrix_analysis.R
\name{activity5_matrix_analysis}
\alias{activity5_matrix_analysis}
\title{Quiz Matrix Analysis}
\usage{
activity5_matrix_analysis(score_matrix, answers, alpha = 0.95, dp = 2)
}
\arguments{
\item{score_matrix}{matrix of scores and best guesses}

\item{answers}{true quiz answers}

\item{alpha}{probability of interval}

\item{dp}{decimal places for score}
}
\value{
list containing
\itemize{
\item scores summary of score per question
\item total_score sum of scores across questions
}
}
\description{
Analyses response for a set of questions in Activity 5 of Playing With Statistics
}
\examples{
team1_score  <- subset(activity5_responses_sm,team == 1)
quiz_score_matrix(team1_score, activity5_quiz_answers) #scores per question and total score for team 1
}
