% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_play_trick.R
\name{activity5_play_trick}
\alias{activity5_play_trick}
\title{Card Trick}
\usage{
activity5_play_trick(
  picture_value = 10,
  time_pause = 1,
  card_path = NULL,
  seed = NULL
)
}
\arguments{
\item{picture_value}{score to give picture cards}

\item{time_pause}{time pause between turning of cards}

\item{card_path}{path for card figures}

\item{seed}{seed fo random number generator}
}
\value{
player's card and prediction
}
\description{
Card trick for Activity 5
}
\examples{
activity5_play_trick()

}
