% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_quiz_answers.R
\docType{data}
\name{activity5_quiz_answers}
\alias{activity5_quiz_answers}
\title{Quiz Answers for Activity 5}
\format{
A data frame with 10 rows and 2 columns
Second column gives answer to respective question
}
\source{
{in-house}
}
\usage{
activity5_quiz_answers
}
\description{
Answers to Quiz Questions in Activity 5 of Playing With Statistics
}
\examples{
activity5_quiz_answers


}
\keyword{datasets}
