% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_response_score.R
\name{activity5_response_score}
\alias{activity5_response_score}
\title{Quiz Score Calculation}
\usage{
activity5_response_score(G, S, Theta, alpha = 0.95, dp = 2)
}
\arguments{
\item{G}{vector of best guesses (one per question)}

\item{S}{vector of measure of accuracy (one per question)}

\item{Theta}{true answer (one per question)}

\item{alpha}{probability of interval}

\item{dp}{decimal places for score}
}
\value{
list containing:
\itemize{
\item scores: score per question
\item total_score: total score across all questions
\item sigma: transformed measure of accuracy (standard deviation) for each question
}
}
\description{
Calculates Score for one or more answers in Activity 5 of Playing With Statistics
}
\examples{
activity5_response_score(c(150, 100), c(20, 60), activity5_quiz_answers$Answer[1:2]) # get scores for first 2 questions

}
