% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_run_app.R
\name{activity5_run_app}
\alias{activity5_run_app}
\title{Run  App 5
Executes app for Activity 5 in Playing With Statistics}
\usage{
activity5_run_app()
}
\description{
Run  App 5
Executes app for Activity 5 in Playing With Statistics
}
\examples{
activity5_run_app()

}
