% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_teams_analysis.R
\name{activity5_teams_analysis}
\alias{activity5_teams_analysis}
\title{Analysis of all team quiz answers}
\usage{
activity5_teams_analysis(response_matrix, answers, alpha = 0.95, dp = 2)
}
\arguments{
\item{response_matrix}{matrix of quiz repsonses}

\item{answers}{true answers}

\item{alpha}{probability of interval}

\item{dp}{decimal places for score}
}
\value{
graphical analysis of score for quiz question
}
\description{
Analyses of all team replies to all questions in Activity 5 of Playing With Statistics
}
\examples{
activity5_teams_analysis(activity5_responses_sm, activity5_quiz_answers) # analysis of answer to question 1 of quiz

}
