% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_execute.R
\name{activity6_execute}
\alias{activity6_execute}
\title{Execute Activity 6}
\usage{
activity6_execute(rand = FALSE, seed = NULL)
}
\arguments{
\item{rand}{randomise settings? (TRUE/FALSE)}

\item{seed}{Value of seed for rnadom number generator}
}
\description{
Individual execution of Activity 6 of Playing With Statistics
}
\examples{
activity6_execute()

}
