% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_player_names.R
\docType{data}
\name{activity6_player_names}
\alias{activity6_player_names}
\title{Player Names}
\format{
A character vecotr of player names
}
\source{
{in-house}
}
\usage{
activity6_player_names
}
\description{
Fictitous player names for Dice Tournament in Activity 6 of Playing With Statistics
}
\examples{
head(activity6_player_names) #list first few names
}
\keyword{datasets}
