% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_player_setup.R
\name{activity6_player_setup}
\alias{activity6_player_setup}
\title{Data entry for Activity 6}
\usage{
activity6_player_setup(nplayers = 32)
}
\arguments{
\item{nplayer}{Number of players in tournament}
}
\value{
character vector of names
}
\description{
Simplifies data entry for Activity 6 in Chapter 6 of Playing With Statistics
}
\examples{
activity6_player_setup(4) # enter data for tournament of 4 players

}
