% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_randomise_players.R
\name{activity6_randomise_players}
\alias{activity6_randomise_players}
\title{Player randomisation for Activity 6}
\usage{
activity6_randomise_players(players, seed = NULL)
}
\arguments{
\item{players}{Character vector of player names}

\item{seed}{value of seed for random number generator}
}
\value{
dataframe with columns:
\itemize{
\item Player: player name
\item Number: randomised player number
}
}
\description{
Facilitates randomisation of players in Activity 6 in Chapter 6 of Playing With Statistics
}
\examples{
activity6_randomise_players(activity6_player_names) # randomise names in activity6_player_names

}
