% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_run_app.R
\name{activity6_run_app}
\alias{activity6_run_app}
\title{Run  App 6
Executes app for Activity 6 in Playing With Statistics}
\usage{
activity6_run_app()
}
\description{
Run  App 6
Executes app for Activity 6 in Playing With Statistics
}
\examples{
activity6_run_app()

}
