% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity6_simulate.R
\name{activity6_simulate}
\alias{activity6_simulate}
\title{Simulate Activity 6}
\usage{
activity6_simulate(
  nrounds = 5,
  n_games_per_match = 5,
  n_rounds_for_estimation = 2,
  nsim = 1000
)
}
\arguments{
\item{nrounds}{number of rounds of tournament}

\item{n_games_per_match}{number of games per match}

\item{n_rounds_for_estimation}{number of rounds to use for data collection prior to estimation}

\item{nsim}{number of matches to simulate when calculating predictions}
}
\value{
value of tournament winner and their win probability at point of estimation
}
\description{
Simulation of Activity 6 of Playing With Statistics
}
\examples{
activity6_simulate()

}
