% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_player_names.R
\docType{data}
\name{activity7_player_names}
\alias{activity7_player_names}
\title{Player Names}
\format{
A character vecotr of player names
}
\source{
{in-house}
}
\usage{
activity7_player_names
}
\description{
Fictitous player names for Dice Tournament in Activity 7 of Playing With Statistics
}
\examples{
head(activity7_player_names) #list first few names
}
\keyword{datasets}
