% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_player_setup.R
\name{activity7_player_setup}
\alias{activity7_player_setup}
\title{Data entry for Activity 7}
\usage{
activity7_player_setup(nplayers = 32)
}
\arguments{
\item{nplayer}{Number of players in tournament}
}
\value{
character vector of names
}
\description{
Simplifies data entry for Activity 7 in Chapter 7 of Playing With Statistics
}
\examples{
activity7_player_setup(4) # enter data for tournament of 4 players

}
