% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_randomise_players.R
\name{activity7_randomise_players}
\alias{activity7_randomise_players}
\title{Player randomisation for Activity 7}
\usage{
activity7_randomise_players(players, seed = NULL)
}
\arguments{
\item{players}{Character vector of player names}

\item{seed}{value of seed for random number generator}
}
\value{
dataframe with columns:
\itemize{
\item Player: player name
\item Number: randomised player number
}
}
\description{
Facilitates randomisation of players in Activity 7 in Chapter 7 of Playing With Statistics
}
\examples{
activity7_randomise_players(player_names) # randomise names in player_names

}
