% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_run_app.R
\name{activity7_run_app}
\alias{activity7_run_app}
\title{Run  App 7
Executes app for Activity 7 in Playing With Statistics}
\usage{
activity7_run_app()
}
\description{
Run  App 7
Executes app for Activity 7 in Playing With Statistics
}
\examples{
activity7_run_app()

}
