% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_run_tournament.R
\name{activity7_run_tournament}
\alias{activity7_run_tournament}
\title{Execute Activity 7}
\usage{
activity7_run_tournament(
  nrounds = 5,
  n_games_per_match = 5,
  n_rounds_for_estimation = 2,
  nsim = 1000,
  seed = NULL
)
}
\arguments{
\item{nrounds}{number of rounds of tournament}

\item{n_games_per_match}{number of games per match}

\item{n_rounds_for_estimation}{number of rounds to use for data collection prior to estimation}

\item{nsim}{number of matches to simulate when calculating predictions}

\item{seed}{value of seed for random number generator}
}
\value{
value of tournament winner and their win probability at point of estimation
}
\description{
Manage dice tournament in Activity 6 of Playing With Statistics
}
\examples{
activity7_run_tournament()

}
