% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_smarties_data.R
\docType{data}
\name{activity7_smarties_data}
\alias{activity7_smarties_data}
\title{Full data collected on Smarties}
\format{
A data frame with 2 rows and 8 columns
Each row corresponds to a pack of Smarties
The columns give the counts of Smarties of each colour per pack
}
\source{
{in-house}
}
\usage{
activity7_smarties_data
}
\description{
Colour frequencies in a dataset of 40 packs of Smarties
}
\examples{
apply(activity7_smarties_data, 1, sum) # counts number of Smarties per pack
}
\keyword{datasets}
