% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_smarties_q1.R
\name{activity7_smarties_q1}
\alias{activity7_smarties_q1}
\title{Analysis of Smarties question 1}
\usage{
activity7_smarties_q1(data_df, nrep = 10000)
}
\arguments{
\item{data_df}{dataframe containing data}

\item{nrep}{number of simulations for calculation of sampling distribution}
}
\value{
histogram of sampling distribution
}
\description{
Analysis of estimate in question 1 of Activity 7 in Playing With Statistics
}
\examples{
activity7_smarties_q1(activity7_smarties_data)

}
