% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity8_run_app.R
\name{activity8_run_app}
\alias{activity8_run_app}
\title{Run Activity 8 app}
\usage{
activity8_run_app(
  commission_max = 0.2,
  bet_rate = 5,
  bet_proportion = 0.1,
  pool_init = 100,
  nominal_stake = 100
)
}
\description{
Run Activity 8 app
}
