% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity8_run_app_sim.R
\name{activity8_run_app_sim}
\alias{activity8_run_app_sim}
\title{Run Horse Race Simulation App
Executes simulation version of horse race app for Activity 8 in Playing With Statistics}
\usage{
activity8_run_app_sim()
}
\description{
Run Horse Race Simulation App
Executes simulation version of horse race app for Activity 8 in Playing With Statistics
}
\examples{
activity8_run_app_sim()

}
