% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_comparison.R
\name{binom_comparison}
\alias{binom_comparison}
\title{Comparsion of binomial distributions}
\usage{
binom_comparison(n = 20, probs = c(0.4, 0.1), nsim = c(100, 1000), seed = NA)
}
\arguments{
\item{n}{Value of n for both binomial distributions}

\item{probs}{Vector of length 2 containing values of p for the 2 binomials}

\item{nsim}{Vector of length 2 specifying sample sizes for simulation}

\item{seed}{Value of seed for random number generator}
}
\value{
Graph of probability functions and bar charts of simulated data
}
\description{
Compares probability distributions and simulated data for 2 binomial distributions
}
\examples{
binom_comparison()
binom_comparison(n = 30, probs = c(0.2, 0.5), nsim= c(25, 250))

}
