% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_norm_approx.R
\name{binom_norm_approx}
\alias{binom_norm_approx}
\title{Binomial-Normal approximation}
\usage{
binom_norm_approx(
  p = 0.2,
  n_min = 5,
  step = 5,
  n = seq(n_min, n_min + 8 * step, by = step),
  ncol = 3,
  scales = "fixed",
  xlim = c(-5, 20),
  approx = TRUE
)
}
\arguments{
\item{p}{in Binomial distribution}

\item{n_min}{smallest value of n}

\item{step}{step size for sequence of n}

\item{n}{values of n in Binomial distribution}

\item{ncol}{number of columns for plotting}

\item{scales}{plotting scales fixed or free?}

\item{xlim}{limit of values of x}

\item{approx}{include Normal approximation (TRUE/FALSE)}
}
\value{
graphical comparison of Binomial and Poisson probability distributions
}
\description{
Compare Binomial and Normal Distributions with identical means as a function of n
}
\examples{
binom_norm_approx(p = 0.2)

}
