% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_pois_approx.R
\name{binom_pois_approx}
\alias{binom_pois_approx}
\title{Binomial-Poisson approximation}
\usage{
binom_pois_approx(
  lambda = 5,
  step = 5,
  n = seq(lambda, lambda + 8 * step, by = step),
  ncol = 3,
  scales = "fixed",
  xlim = c(0, 3 * lambda),
  approx = TRUE
)
}
\arguments{
\item{lambda}{mean of Poisson distribution}

\item{step}{step size for sequence of n}

\item{n}{values of n in Binomial distribution}

\item{ncol}{number of columns for plotting}

\item{scales}{plotting scales fixed or free?}

\item{xlim}{limit of values of x}

\item{approx}{include poisson approximation (TRUE/FALSE)}
}
\value{
graphical comparison of Binomial and Poisson probability distributions
}
\description{
Compare Binomial and Poisson Distributions with Identical Means as function of n
}
\examples{
binom_pois_approx(lambda=5)

}
