% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dredge.R
\name{data_dredge}
\alias{data_dredge}
\title{Illustration of Effects of Data Dredging}
\usage{
data_dredge(n_data = 100, n_var = 50, sig_x = 10, sig_y = 5, seed = 3)
}
\arguments{
\item{n_data}{number of rows of data}

\item{sig_x}{standard deviation of y values}

\item{seed}{seed to initialise randomisation}

\item{n_games_per_match}{number of variables}
}
\value{
summary of most significant linear relationship between y and any of the x variables
}
\description{
Data dredging example discussed in Chapter 7 of Playing With Statistics
}
\examples{
data_dredge()

}
