% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_choice.R
\name{estimate_choice}
\alias{estimate_choice}
\title{Estimate choice exercise}
\usage{
estimate_choice(
  theta = 5,
  m1 = 5,
  sd1 = 1.5,
  m2 = 4.75,
  sd2 = 0.25,
  rho = 0.75
)
}
\arguments{
\item{theta}{value of parameter to be estimated}

\item{m1}{mean of first estimator}

\item{sd1}{standard deviation of first estimator}

\item{m2}{mean of second estimator}

\item{sd2}{standard deviation of second estimator}

\item{xG_mod}{fitted xG model}
}
\value{
pair of estimates
}
\description{
Provides 2 estimates of a quantity as required in Chapter 5 of Playing With Statistics
}
\examples{
estimate_choice()

}
