% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPrivate.R
\name{.getAPI}
\alias{.getAPI}
\title{A function to get API keys from the website \url{http://www.wunderground.com}.}
\arguments{
\item{.getAPI}{returns an API key from the website \url{http://www.wunderground.com}}
}
\description{
A function to get API keys from the website \url{http://www.wunderground.com}.
}

