\name{getHist-methods}
\docType{methods}
\alias{getHist-methods}
\alias{getHist,myWeatherObj-method}
\title{ Method for obtaining weather history for a weather object  getHist  ~~}
\description{
 Method for obtaining weather history for a weather object \code{getHist}  ~~
}
\section{Methods}{
\describe{

\item{\code{signature(object = "myWeatherObj")}}{
Given a weather object this allows the user to specify a date range
and then use that to obtain weather history data for all stations in
that object. It returns a "weather history object" that can then be mined
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
