% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPWS.R
\name{getPWS}
\alias{getPWS}
\title{get an object of class "PWS"}
\arguments{
\item{lat}{a numeric that specifies the latitude of the original search location; must be between -180 and 180}

\item{lon}{a numeric that specifies the longitude of the original search location; must be between -180 and 180}

\item{zipcode}{a numeric that specifies the zipcode of the original search location; only valid for locations within USA}

\item{city}{a single character string that specifies the city of the original search location}

\item{state}{a single character string that specifies the state of the original search location; should be supplied if the location is within USA}

\item{country}{a single character string that specifies the country of the original search location; should be supplied if the location is outside USA}

\item{distance}{a positive numeric which specifies the distance range of the nearby PWS from the original search location}
}
\value{
An S4 object of class \code{\link{PWS-class}} containing the information about the original search location and the nearby PWS. The object contains the following slots:
\describe{
\item{\code{city}}{a single character string that shows the city of the original search location}
\item{\code{state}}{a single character string that shows the state of the original search location; is character(0) if the origin is outside the U.S.A}
\item{\code{country}}{a single character string that shows the country of the original search location; is character(0) if the the origin is within the U.S.A}
\item{\code{lat}}{a numeric that shows the latitude of the original search location; must be between -180 and 180}
\item{\code{lon}}{a numeric that shows the longitude of the original search location; must be between -180 and 180}
\item{\code{data}}{a data frame that contains the information of the nearby PWS; in particular, the columns of the data frame should be: \describe{
\item{\code{neighborhood}}{a single character string that shows the neighborhood of the PWS}
\item{\code{city}}{a single character string that shows the city of the PWS}
\item{\code{state}}{a single character string that shows the state of the PWS}
\item{\code{country}}{a single character string that shows the country of the PWS}
\item{\code{id}}{a numeric that shows the id of the PWS}
\item{\code{lat}}{a numeric that shows the latitude of the PWS}
\item{\code{lon}}{a numeric that shows the longitude of the PWS}
\item{\code{distance_km}}{a positive integer that shows the distance from the PWS to the original search location in kilometers}
\item{\code{distance_mi}}{a positive integer that shows the distance from the PWS to the original search location in miles}
}}}
see \code{\link{PWS-class}} for more details
}
\description{
Gets the nearby PWS (personal weather station) from \url{http://www.wunderground.com} based on a original search location and a distance range.
}
\details{
Three methods of specifying a original search location are defined. The function first searches the location by \code{lat} and \code{lon} if they are supplied. If \code{lat} and \code{lon} are not supplied, the function next searches the location by the \code{zipcode} if it is supplied. If again \code{zipcode} is not supplied, then the function searches the location by the \code{city}. If \code{state} is supplied, then the city will be searched within the USA. Otherwise if \code{country} is supplied, then the city will be searched outside the USA.
The function then gets the nearby PWS of the original search location within a distance range which is specified by \code{distance}.
}
\examples{
a = getPWS(zipcode = 94305)
b = getPWS(city = "San Francisco", state = "CA")
c = getPWS(lat = 37, lon = -120)

}
\seealso{
\code{\link{PWS-class}}
}

