% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPWSweather.R
\name{getPWSweather}
\alias{getPWSweather}
\title{get an object of class "PWSweather"}
\usage{
getPWSweather(pws, start, end)
}
\arguments{
\item{pws}{a PWS object that contains the information about the original search location and the nearby PWS}

\item{start}{a single character string in HH:MM or H:MM form that shows the starting time of the time interval}

\item{end}{a single character string in HH:MM or H:MM form that shows the ending time of the time interval; must be greater than starting time}
}
\value{
An S4 object of class \code{\link{PWSweather-class}} contains the information about the weather condition of some PWS (personal weather stations) on a given time interval
}
\description{
Gets the up to date weather information about nearby PWS (personal weather station) from \url{http://www.wunderground.com} based on a PWS object, a starting time and an ending time
}
\details{
The function retrieves weather information about all stations listed in \code{pws} and subset the data based on the time interval which is specified by \code{start} and \code{end}
}
\note{
please use this function with care since there are limits to the number of API calls per minute and per day
}
\examples{
data(offlineExample)
a = getPWSweather(offlineExample, "3:00", "17:00")
validObject(a)

}
\seealso{
\code{\link{PWSweather-class}}
}

