\name{getWobjData-methods}
\docType{methods}
\alias{getWobjData-methods}
\alias{getWobjData,myWeatherObj-method}
\title{ Accessor Method for PWS Weather Objects }
\description{
 Accessor Methods for function \code{getWobjData}  ~~
}
\section{Methods}{
\describe{

\item{\code{signature(object = "myWeatherObj")}}{
This is an accessor function for PWS weather objects. In particular
it allows the user to access columns/attributes of the data frame
in the Weather Object.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
