% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_race_results.R
\docType{data}
\name{horse_race_results}
\alias{horse_race_results}
\title{Horse race results for Activity 8}
\format{
A data frame with 50 rows and 7 columns
}
\source{
{in-house}
}
\usage{
horse_race_results
}
\description{
Horse race results for Activity 8 of Playing With Statistics
}
\examples{
horse_race_results


}
\keyword{datasets}
