% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputePWS.R, R/plotPWSweather.R
\docType{methods}
\name{imputePWS}
\alias{imputePWS}
\alias{imputePWS,PWSweather-method}
\alias{plotPWSweather,PWSweather-method}
\title{register the \code{data} slot in an object of class \code{\link{PWSweather-class}} for some reported \code{variable} from individual stations}
\usage{
\S4method{imputePWS}{PWSweather}(object, variable)

\S4method{plotPWSweather}{PWSweather}(object, variable)
}
\arguments{
\item{object}{an object of class \code{\link{PWSweather-class}}}

\item{variable}{a single character string that target variable that needs to be imputed, must be one of "tempi" (Temperature), "hum" (Humidity), "wspdi" (Windspeed) or "pressurei" (Pressure)}
}
\value{
a data frame contain columns of imputed value of each station, and a column of timestamp
}
\description{
\code{imputePWS} registers the \code{data} slot in an object of class \code{\link{PWS-class}} for some reported \code{variable} by interpolating or otherwise approximating the variable for each station through linear approximation
}
\section{Methods (by class)}{
\itemize{
\item \code{PWSweather}: imputes an object of class "PWSweather" by a given variable

\item \code{PWSweather}: imputes an object of class "PWSweather" by a given variable
}}
\examples{
data(offlineWeather)
d = imputePWS(offlineWeather, 'tempi')

}
\seealso{
\code{\link{PWSweather-class}}
}
\keyword{methods}

