% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPWS.R
\name{isPWS}
\alias{isPWS}
\title{check whether an object is a "PWS" class}
\usage{
isPWS(object)
}
\arguments{
\item{object}{an object could be any class}
}
\value{
a binary response is either TRUE or FALSE
}
\description{
check whether an object is a "PWS" class
}
\examples{
data(offlineExample)
isPWS(offlineExample)
}
\seealso{
\code{\link{PWS-class}}
}

