% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPWSweather.R
\name{isPWSweather}
\alias{isPWSweather}
\title{check whether an object is a "PWSweather" class}
\usage{
isPWSweather(object)
}
\arguments{
\item{object}{an object could be any class}
}
\value{
a binary response is either TRUE or FALSE
}
\description{
check whether an object is a "PWSweather" class
}
\examples{
data(offlineWeather)
isPWSweather(offlineWeather)
}
\seealso{
\code{\link{PWSweather-class}}
}

