% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_activity_safe.R
\name{launch_activity}
\alias{launch_activity}
\title{Launch a Shiny activity in a background R process}
\usage{
launch_activity(app_name, timeout = 15)
}
\arguments{
\item{app_name}{Name of app directory under inst/shiny/}

\item{timeout}{Maximum startup wait time in seconds}
}
\value{
Invisibly returns the callr process object
}
\description{
Launch a Shiny activity in a background R process
}
