% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_activity_from_lab.R
\name{launch_activity_from_lab}
\alias{launch_activity_from_lab}
\title{Launch activity for use from the Statistics Lab}
\usage{
launch_activity_from_lab(chapter)
}
\arguments{
\item{chapter}{Chapter number}
}
\description{
Starts the activity app and waits briefly for startup.
}
\keyword{internal}
