% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_win_probs.R
\name{match_win_probs}
\alias{match_win_probs}
\title{Match Win Probabilities}
\usage{
match_win_probs(team1_pars, team2_pars, home_adv_par, score_max = 20, sig = 4)
}
\arguments{
\item{team1_pars}{vector of attack and defence parameters for team 1}

\item{team2_pars}{vector of attack and defence parameters for team 2}

\item{home_adv_par}{home advantage parameter}

\item{score_max}{upper limit for either team goal count}

\item{sig}{significant figures in output}
}
\value{
vector of 1/x/2 probabilities
}
\description{
Match win probability calculation based on standard model with independent Poisson means for goal scores per team
}
\examples{
match_win_probs(c(.346,.192), c(.521,.249), .063) # man city vs liverpool example from Chapter 8

}
