% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_goals_sim.R
\name{multi_goals_sim}
\alias{multi_goals_sim}
\title{Simulate goals in multiple games}
\usage{
multi_goals_sim(
  n_sim,
  pois_mean = 100,
  beta_alpha_1 = 0.01,
  beta_alpha_2 = 0.5,
  plot_max = NULL,
  seed = NULL
)
}
\arguments{
\item{n_sim}{Number of games to simulate}

\item{pois_mean}{Average number of events per game}

\item{beta_alpha_1}{First parameter that affects probability events become goals}

\item{beta_alpha_2}{Second parameter that affects probability events become goals}

\item{plot_max}{Maximum value for subsequent plotting (optional)}

\item{seed}{Set seed to enable identical simulation across calls}
}
\value{
list containing number of simulated goals per game and a frequency table
}
\description{
Simulates goals in multiple games as described in Chapter 1
}
\examples{
multi_goals_sim(n_sim = 1000)

}
