% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_trick.R
\name{multi_trick}
\alias{multi_trick}
\title{Multi Trick}
\usage{
multi_trick(nrep = 10000, picture_value = 10)
}
\arguments{
\item{nrep}{number of repetitions}

\item{picture_value}{score to give picture cards}
}
\value{
summary statistics from multiple trick runs
}
\description{
Multiple runs of trick for Activity 5
}
\examples{
multi_trick()

}
