% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL24_pars.R
\docType{data}
\name{PL24_pars}
\alias{PL24_pars}
\title{Premier League Model Parameters}
\format{
List with 2 objects:
1: value of tau, home advantage parameter
2: dataframe of team parameters (alpha, beta)
}
\source{
{in-house}
}
\usage{
PL24_pars
}
\description{
Premier League Model Parameters
}
\examples{
PL24_pars[[1]] # home advantage parameter
with(PL24_pars[[2]], plot(alpha, beta)) # plot of defence against attack parameters

}
\keyword{datasets}
