% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL25_pars.R
\docType{data}
\name{PL25_pars}
\alias{PL25_pars}
\title{Premier League Model Parameters (25/26 season)}
\format{
List with 2 objects:
1: value of tau, home advantage parameter
2: dataframe of team parameters (alpha, beta)
}
\source{
{in-house}
}
\usage{
PL25_pars
}
\description{
Premier League Model Parameters (25/26 season)
}
\examples{
PL25_pars[[1]] # home advantage parameter
with(PL25_pars[[2]], plot(alpha, beta)) # plot of defence against attack parameters

}
\keyword{datasets}
