% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL25_schedule.R
\docType{data}
\name{PL25_schedule}
\alias{PL25_schedule}
\title{Premier League Scedule 25-26}
\format{
dataframe of match schedules by round
}
\source{
{in-house}
}
\usage{
PL25_schedule
}
\description{
Premier League Scedule 25-26
}
\examples{
head(PL25_schedule) # home advantage parameter

}
\keyword{datasets}
