% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL26_schedule.R
\docType{data}
\name{PL26_schedule}
\alias{PL26_schedule}
\title{Premier League Schedule 26-27}
\format{
dataframe of match schedules by round
}
\source{
{in-house}
}
\usage{
PL26_schedule
}
\description{
Premier League Schedule 26-27
}
\examples{
head(PL26_schedule) # home advantage parameter

}
\keyword{datasets}
