% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_cards_data.R
\docType{data}
\name{PL_cards}
\alias{PL_cards}
\title{Premier League Red Cards  per Season}
\format{
A data frame with 33 rows and 4 columns
Each row corresponds to one season of Premier League
The columns give season, number of red cards, number of games, and rate of red cards per game
}
\usage{
PL_cards
}
\description{
Red cards per season in Premier League season 1992/93 to 2024/25
}
\examples{
mean(PL_cards$rate1) # Mean number of red cards per game per season
}
\keyword{datasets}
