% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_goals_data.R
\docType{data}
\name{PL_goals}
\alias{PL_goals}
\title{Premier League Total Goals per Game}
\format{
A data frame with 12406 rows and 3 columns
Each row corresponds to a Premier League fixture
The columns give date, season and total goals
}
\source{
{https://www.football-data.co.uk/data.php}
}
\usage{
PL_goals
}
\description{
Total Goals scored in Premier League fixtures from season 1993/94 to 2023/24
}
\examples{

mean(PL_goals$goals) # Find mean number of goals over all fixtures

mean(subset(PL_goals$goals, PL_goals$season == "2023-24")) # Find mean number of goals over fixtures  2023/24 season

hist(PL_goals$goals, col="lightblue", xlab= "Number of Goals", main = "Histogram of goals per game in Premier League Matches") # histogram of all goals data

}
\keyword{datasets}
