% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_predicted_goals_data.R
\docType{data}
\name{PL_predicted_goals}
\alias{PL_predicted_goals}
\title{Predicted GOals in Premier League Fixtures}
\format{
A data frame with 6302 rows and 3 columns
Each row corresponds to a Premier League fixture
The columns give fixture number, predicted total goals and observed total goals
}
\source{
in-house
}
\usage{
PL_predicted_goals
}
\description{
Predicted and observed goals per fixture in sereis of Premier League fixtures
}
\examples{

with(PL_predicted_goals, plot(predicted_goals, jitter(observed_goals, .25), pch=16), ylab = "predicted_goals") # plot of observed against predicted goals (jittered to avoid overlap)
}
\keyword{datasets}
