% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_results.R
\docType{data}
\name{PL_results}
\alias{PL_results}
\title{Premier League Results to season 25-26}
\format{
dataframe of match results
}
\source{
{in-house}
}
\usage{
PL_results
}
\description{
Premier League Results to 25-26
}
\examples{
head(PL_results) #

}
\keyword{datasets}
