% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_names.R
\docType{data}
\name{player_names}
\alias{player_names}
\title{Player Names}
\format{
A character vecotr of player names
}
\source{
{in-house}
}
\usage{
player_names
}
\description{
Fictitous player names for Dice Tournament in Activity 7 of Playing With Statistics
}
\examples{
head(player_names) #list first few names
}
\keyword{datasets}
