% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bp.R
\name{plot_bp}
\alias{plot_bp}
\title{THe Birthday Problem}
\usage{
plot_bp(nmax = 60, p_target = 0.5)
}
\arguments{
\item{nmax}{upper value for plotting}
}
\value{
graph of probabiltiys for the birthday problem
}
\description{
Probabiltiy calculations for the classical birthday problem
}
\examples{
prop._bp()

}
