% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPWS.R
\docType{methods}
\name{plotPWS}
\alias{plotPWS}
\alias{plotPWS,PWS-method}
\title{plot the geographical location of the object of class \code{\link{PWS-class}}}
\usage{
\S4method{plotPWS}{PWS}(object, googleMap = TRUE)
}
\arguments{
\item{object}{an object of class \code{\link{PWS-class}}}

\item{googleMap}{logical variable that indicates whether google map should be used for the plotting}
}
\description{
\code{plotPWS} plots the original search location and the nearby PWS's that are contained in an object of class \code{\link{PWS-class}}. It gives a visualization of the relative locations of origin and PWS. Users can choose to plot either over google map or as plain points.
}
\section{Methods (by class)}{
\itemize{
\item \code{PWS}: plot the geographical location based on an object of class "PWS"
}}
\examples{
data(offlineExample)
offlineExample
plotPWS(offlineExample, googleMap = FALSE) #plot without getting online google map
plotPWS(offlineExample, googleMap = TRUE) #plot using online google map

}
\seealso{
\code{\link{PWS-class}}
}
\keyword{graphics}
\keyword{methods}

