% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPWScurrent.R
\docType{methods}
\name{plotPWScurrent}
\alias{plotPWScurrent}
\alias{plotPWScurrent,PWS-method}
\title{plot the current weather condition based on the object of class \code{\link{PWS-class}}}
\usage{
\S4method{plotPWScurrent}{PWS}(object, googleMap = TRUE)
}
\arguments{
\item{object}{an object of class \code{\link{PWS-class}}}

\item{googleMap}{logical variable that indicates whether google map should be used for the plotting}
}
\description{
\code{plotPWS} plots the current weather conditions for original search location and the nearby PWS's that are contained in an object of class \code{\link{PWS-class}}. It gets the weather condition by API calls from the website \url{http://www.wunderground.com}. The labels on the graph show temperatures in Fahrenheit scale. Users can choose to plot either over google map or as plain points.
}
\section{Methods (by class)}{
\itemize{
\item \code{PWS}: plot the current weather condition based on an object of class "PWS"
}}
\note{
Please use this function with care since there are limits to the number of API calls per minute and per day from the website \url{http://www.wunderground.com}.
}
\examples{
data(offlineExample)
offlineExample
plotPWScurrent(offlineExample, googleMap = FALSE) #plot without getting online google map
plotPWScurrent(offlineExample, googleMap = TRUE) #plot using online google map


}
\seealso{
\code{\link{PWS-class}}
}
\keyword{graphics}
\keyword{methods}

