% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPWSweather.R
\name{plotPWSweather}
\alias{plotPWSweather}
\title{visualize properties of the \code{data} slot in an object of class \code{\link{PWSweather-class}}}
\arguments{
\item{object}{an object of class \code{\link{PWSweather-class}}}

\item{variable}{a single character string that represent the target of the plot, must be one of "Temperature", "Humidity", "Windspeed" or "Pressure"}
}
\value{
a plot shows the attribute define by \code{variable} vs time of each station
}
\description{
\code{plotPWSweather} plots specific \code{variable} of each station in \code{data} slot in an object of class \code{\link{PWSweather-class}}
}
\examples{
data(offlineWeather)
plotPWSweather(offlineWeather, 'Temperature')
plotPWSweather(offlineWeather, 'Humidity')

}
\seealso{
\code{\link{PWSweather-class}}
}
\keyword{graphics}
\keyword{methods}

