% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.diff.R
\name{prop.diff}
\alias{prop.diff}
\title{Test for difference in proportions}
\usage{
prop.diff(
  counts = c(21, 9),
  trials = c(31, 23),
  alpha = 0.95,
  nsim = 10000,
  seed = NULL,
  ci_lim = NULL
)
}
\arguments{
\item{counts}{vector of counts in 2 groups}

\item{trials}{number of trials in each group}

\item{alpha}{level of confidence}

\item{nsim}{number of simulations}

\item{seed}{Set seed to enable identical simulation across calls}

\item{ci_lim}{vector of limits for plotting confidence interval}
}
\value{
standard error of difference and alpha-level confidence intervals (plots as side effect)
}
\description{
Tests for difference in proportions of two samples
}
\examples{
prop.diff()

}
