% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{PWS-class}
\alias{PWS-class}
\alias{getCity,PWS-method}
\alias{getCountry,PWS-method}
\alias{getData,PWS-method}
\alias{getLat,PWS-method}
\alias{getLon,PWS-method}
\alias{getState,PWS-method}
\title{An S4 class "PWS"}
\description{
The class of object returned by the \code{\link{getPWS}} function. The PWS class represents the original search location and its nearby PWS (personal weather stations), which are retrieved from the website \url{http://www.wunderground.com}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getCity}: returns the city of the original search location

\item \code{getState}: returns the state of the original search location

\item \code{getCountry}: returns the country of the original search location

\item \code{getLat}: returns the latitude of the original search location

\item \code{getLon}: returns the longitude of the original search location

\item \code{getData}: returns the data of the original search location
}}
\section{Slots}{

\describe{
\item{\code{city}}{a single character string that shows the city of the original search location}

\item{\code{state}}{a single character string that shows the state of the original search location; is character(0) if the origin is outside the U.S.A}

\item{\code{country}}{a single character string that shows the country of the original search location; is character(0) if the the origin is within the U.S.A}

\item{\code{lat}}{a numeric that shows the latitude of the original search location; must be between -180 and 180}

\item{\code{lon}}{a numeric that shows the longitude of the original search location; must be between -180 and 180}

\item{\code{data}}{a data frame that contains the information of the nearby PWS; in particular, the columns of the data frame should be: \describe{
\item{\code{neighborhood}}{a single character string that shows the neighborhood of the PWS}
\item{\code{city}}{a single character string that shows the city of the PWS}
\item{\code{state}}{a single character string that shows the state of the PWS}
\item{\code{country}}{a single character string that shows the country of the PWS}
\item{\code{id}}{a numeric that shows the id of the PWS}
\item{\code{lat}}{a numeric that shows the latitude of the PWS}
\item{\code{lon}}{a numeric that shows the longitude of the PWS}
\item{\code{distance_km}}{a positive integer that shows the distance from the PWS to the original search location in kilometers}
\item{\code{distance_mi}}{a positive integer that shows the distance from the PWS to the original search location in miles}
}}
}}
\note{
The class \code{\link{PWS-class}} has a validity (see \code{\link{validObject}}) check function.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form new("PWS", ...) or by calling the function \code{\link{getPWS}}.
}
\examples{
a = getPWS(zipcode = 94305)
b = getPWS(city = "San Francisco", state = "CA")
validObject(a)

}
\seealso{
\code{\link{getPWS}}
}
\keyword{classes}

