% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pws-package.R
\docType{package}
\name{pws-package}
\alias{pws}
\alias{pws-package}
\title{pws: Playing With Statistics}
\description{
Description: Tools and data for Playing With Statistics. Includes match data derived from the engsoccerdata package.
}
\author{
\strong{Maintainer}: Stuart Coles \email{stuart.coles1111@gmail.com}

}
\keyword{internal}
