% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWS.R
\docType{package}
\name{PWS}
\alias{PWS}
\alias{PWS-package}
\title{A project package with R}
\description{
This package extract, organize, visualize and predict local weather data.
User can use this package to compare data at nearby stations to show microclimate patterns.
See examples for functions \code{\link{PWSweather-class}}, \code{\link{PWS-class}}, \code{\link{subsetPWS}}, 
\code{\link{plotPWS}}, \code{\link{plotPWScurrent}}, \code{\link{getPWS}}, \code{\link{getPWSweather}},
\code{\link{imputePWS}}, \code{\link{plotPWSweather}}
}
\references{
Wunderground project at \url{http://www.wunderground.com/}
}
\seealso{
\code{\link{PWS-class}}

\code{\link{PWSweather-class}}

\code{\link{subsetPWS}}

\code{\link{plotPWS}}

\code{\link{plotPWScurrent}}

\code{\link{getPWS}}

\code{\link{getPWSweather}}

\code{\link{imputePWS}}

\code{\link{plotPWSweather}}
}

