% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{PWSweather-class}
\alias{PWSweather-class}
\alias{getEndTime,PWSweather-method}
\alias{getStartTime,PWSweather-method}
\alias{getWeatherData,PWSweather-method}
\title{An S4 class "PWSweather"}
\description{
The class of object returned by the \code{\link{getPWSweather}} function. The PWSweather class represents the weather condition of some PWS (personal weather stations) on a given time interval. The data is retrieved from the website \url{http://www.wunderground.com}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getWeatherData}: returns the weather data of the nearby stations

\item \code{getStartTime}: returns the starting time of the time interval

\item \code{getEndTime}: returns the ending time of the time interval
}}
\section{Slots}{

\describe{
\item{\code{start}}{a single character string in HH:MM or H:MM form that shows the starting time of the time interval}

\item{\code{end}}{a single character string in HH:MM or H:MM form that shows the ending time of the time interval; must be greater than starting time}

\item{\code{data}}{a list of data frame that contains the weather of the nearby PWS}
}}
\note{
The class \code{\link{PWSweather-class}} has a validity (see \code{\link{validObject}}) check function.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form new("PWSweather", ...) or by calling the function \code{\link{getPWSweather}}.
}
\examples{
data(offlineExample)
a = getPWSweather(offlineExample, "3:00", "17:00")
validObject(a)

}
\seealso{
\code{\link{getPWSweather}}
}
\keyword{classes}

