% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_activity.R, R/run_activity_safe.R
\name{run_activity}
\alias{run_activity}
\title{Run an activity app}
\usage{
run_activity(chapter)

run_activity(chapter)
}
\arguments{
\item{chapter}{Chapter number}
}
\description{
Run an activity app

Run an activity app
}
