% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_trick.R
\name{single_trick}
\alias{single_trick}
\title{Single Trick}
\usage{
single_trick(picture_value = 10, show_result = TRUE, seed = NULL)
}
\arguments{
\item{picture_value}{score to give picture cards}

\item{show_result}{should result be displayed? (TRUE/FALSE)}

\item{seed}{value of seed for random number generator}
}
\value{
player and magician final card
}
\description{
Output of single trick simulation for Activity 5
}
\examples{
single_trick()

}
