% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPWS.R
\docType{methods}
\name{subsetPWS}
\alias{subsetPWS}
\alias{subsetPWS,PWS-method}
\title{subset the \code{data} slot in an object of class \code{\link{PWS-class}}}
\usage{
\S4method{subsetPWS}{PWS}(object, distance)
}
\arguments{
\item{object}{an object of class \code{\link{PWS-class}}}

\item{distance}{a positive numeric that specifies the distance (in kilometers) to subset the \code{data} slot}
}
\value{
an object of class \code{\link{PWS-class}}
}
\description{
\code{subsetPWS} extracts subtables of the \code{data} slot in an object of class \code{\link{PWS-class}} by distance (with respect to the original search location). It keeps the PWS which are within the specified distance of original search location.
}
\section{Methods (by class)}{
\itemize{
\item \code{PWS}: subsets an object of class "PWS"
}}
\examples{
data(offlineExample)
offlineExample
b = subsetPWS(offlineExample, 1)

}
\seealso{
\code{\link{PWS-class}}
}
\keyword{methods}

