% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trick_fail_prob_se.R
\name{trick_fail_prob_se}
\alias{trick_fail_prob_se}
\title{Trick failure probability SE}
\usage{
trick_fail_prob_se(trick_res, n_resample = 10000, seed = NULL)
}
\arguments{
\item{trick_res}{output from multi_trick}

\item{n_resample}{number of resamples}

\item{seed}{value of seed for random number generator}
}
\value{
value of standard error
}
\description{
Standard error calculation of trick failure probability for Activity 5
}
\examples{
trick_res = multi_trick()
          trick_fail_prob_se(trick_res)

}
