% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{.validPWS}
\alias{.validPWS}
\title{A validation check for S4 class "PWS"}
\usage{
.validPWS(object)
}
\arguments{
\item{object}{an object of class "PWS"}
}
\description{
A validation check for S4 class "PWS"
}

