% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{.validPWSweather}
\alias{.validPWSweather}
\title{A validation check for S4 class "PWSweather"}
\usage{
.validPWSweather(object)
}
\arguments{
\item{object}{an object of class "PWSweather"}
}
\description{
A validation check for S4 class "PWSweather"
}

