% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGpred.R
\name{xGpred}
\alias{xGpred}
\title{Predict xG}
\usage{
xGpred(xG_mod, x, y, body)
}
\arguments{
\item{xG_mod}{fitted xG model}

\item{x}{vector of x-coordinates of events}

\item{y}{vector of y-coordinates of events}

\item{body}{character vector of elements ("Foot"/"Head") indicating type of shot per event}
}
\value{
predicted xG and event info
}
\description{
Predicts xG according to model with specified parameters
}
\examples{
xG_data <- xGsim()
xG_model <- xGfit(xG_data)
xGpred(xG_model, x= c(5, 15, -20, 7, -12),
        y = c(10, 18, 6, 14, 25),
        body = c("Head", "Head",
            "Foot", "Foot", "Foot"))

}
