% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGsim.R
\name{xGsim}
\alias{xGsim}
\title{Simulate xG}
\usage{
xGsim(
  n_data,
  bodypar = 0.2,
  distpar = 0.2,
  anglepar = 3,
  intercept = 0.25,
  dist_coeff = -0.1,
  dist_body_inter = -0.25,
  body_coeff = -0.05,
  angle_coeff = -0.8,
  seed = NULL
)
}
\arguments{
\item{n_data}{Number of events to simulate}

\item{bodypar}{parameter determining rate of chances with Head/Foot respectively}

\item{distpar}{paramter determining distribution of distances of shots}

\item{anglepar}{paramter determining distribution of angles of shots}

\item{intercept}{intercept in linear predictor part of model}

\item{dist_coeff}{coefficient of distance in linear predictor part of model}

\item{dist_body_inter}{interaction term for body type and distance in linear predictor part of model}

\item{body_coeff}{coefficient of body type in linear predictor part of model}

\item{angle_coeff}{coefficient of angle in linear predictor part of model}
}
\value{
List comprising
data: the simulated data
pars: the model parameters used in simulating the model
}
\description{
Simulates xG according to model with specified parameters from a specified number of scoring opportunities
}
\examples{
xGsim()

}
