% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity4_response_analysis.R
\name{activity4_response_analysis}
\alias{activity4_response_analysis}
\title{Quiz Answer Analysis}
\usage{
activity4_response_analysis(
  G,
  S,
  Theta,
  alpha = 0.95,
  lines = TRUE,
  final_score_only = FALSE,
  dp = 2
)
}
\arguments{
\item{G}{best guess}

\item{S}{measure of accuracy}

\item{Theta}{true answer (if data not NULL, just give index of dataset)}

\item{alpha}{probability of interval}

\item{lines}{add lines to plots}

\item{dp}{decimal places for score}

\item{final}{show plot for final score only (TRUE/FALSE)}

\item{data}{dataset (if any) containing answers}
}
\value{
graphical analysis of score for quiz question
}
\description{
Analyses response for a single question in Activity 4 of Playing With Statistics
}
\examples{
activity4_response_analysis(150, 20, activity4_quiz_answers$Answer[1], lines = T) # analysis of answer to question 1 of quiz

}
