% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trick_fail_prob_se.R
\name{activity5_prob_se}
\alias{activity5_prob_se}
\title{Trick failure probability SE}
\usage{
activity5_prob_se(trick_res, n_resample = 1000, seed = NULL)
}
\arguments{
\item{trick_res}{output from multi_trick}

\item{n_resample}{number of resamples}

\item{seed}{value of seed for random number generator}
}
\value{
value of standard error
}
\description{
Standard error calculation of trick failure probability for Activity 5
}
\examples{
trick_res <- activity5_trick_analysis()
          activity5_prob_se(trick_res)

}
