% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_response_analysis.R
\name{activity5_response_analysis}
\alias{activity5_response_analysis}
\title{Quiz Answer Analysis}
\usage{
activity5_response_analysis(
  G,
  S,
  Theta,
  alpha = 0.95,
  lines = FALSE,
  final_score_only = FALSE,
  dp = 2
)
}
\arguments{
\item{G}{best guess}

\item{S}{measure of accuracy}

\item{Theta}{true answer}

\item{alpha}{probability of interval}

\item{lines}{add lines to plots}

\item{dp}{decimal places for score}

\item{final}{show plot for final score only (TRUE/FALSE)}
}
\value{
graphical analysis of score for quiz question
}
\description{
Analyses response for a single question in Activity 5 of Playing With Statistics
}
\examples{
activity5_response_analysis(150, 20, activity5_quiz_answers$Answer[1], lines = T) # analysis of answer to question 1 of quiz

}
