% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity5_trick_analysis.R
\name{activity5_trick_analysis}
\alias{activity5_trick_analysis}
\title{Analysis of card trick performance for Activity 5}
\usage{
activity5_trick_analysis(nrep = 10000, picture_value = 10, seed = NULL)
}
\arguments{
\item{nrep}{number of repetitions}

\item{picture_value}{score to give picture cards}

\item{seed}{value of seed for random number generator}
}
\value{
summary statistics from multiple trick runs
}
\description{
Multiple runs of trick for Activity 5
}
\examples{
activity5_trick_analysis()

}
