% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_execute.R
\name{activity7_execute}
\alias{activity7_execute}
\title{Execute Activity 7}
\usage{
activity7_execute(rand = FALSE, seed = NULL)
}
\arguments{
\item{rand}{randomise settings? (TRUE/FALSE)}

\item{seed}{value of seed for rnadom number generator}
}
\description{
Individual execution of Activity 7 of Playing With Statistics
}
\examples{
activity7_execute()

}
