% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity7_execute.R
\name{activity7_execute}
\alias{activity7_execute}
\title{Execute Activity 7}
\usage{
activity7_execute(weight = c(4, 2, 3), mu = 193, sd = 10, rand = FALSE)
}
\arguments{
\item{weight}{weights for spend effects on different categories}

\item{mu}{baseline distance}

\item{sd}{standard deviation of distances}

\item{rand}{randomise settings? (TRUE/FALSE)}
}
\description{
Individual execution of Activity 7 of Playing With Statistics
}
\examples{
activity7_execute()

}
