% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity8_run_app.R
\name{activity8_run_app}
\alias{activity8_run_app}
\title{Run Horse Race App}
\usage{
activity8_run_app(mode = c("multiplayer", "singleplayer"))
}
\arguments{
\item{mode}{"multiplayer" or "singleplayer"}
}
\description{
Run Horse Race App
}
