% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{activity8_run_app}
\alias{activity8_run_app}
\title{Run Horse Race App
Executes horse race app for Activity 8 in Playing With Statistics}
\usage{
activity8_run_app()
}
\description{
Run Horse Race App
Executes horse race app for Activity 8 in Playing With Statistics
}
\examples{
activity8_run_app()

}
