% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_mean_series.R
\name{dice_mean_series}
\alias{dice_mean_series}
\title{Analyse running means from sequences of dice throws}
\usage{
dice_mean_series(n_rolls = 1000, nrep = 9, ncol = 3)
}
\arguments{
\item{n_rolls}{Maximum number of dice rolls}

\item{nrep}{Number of repetitions for each experiment}

\item{ncol}{Number of columns for plots}
}
\value{
Series of cumulative running means of dice scores in each experiment
}
\description{
Shows how sample means of large samples converge based on mean score of standard dice rolls
}
\examples{
dice_mean_series()

}
