% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals_sim.R
\name{goals_sim}
\alias{goals_sim}
\title{Simulate goals}
\usage{
goals_sim(
  x,
  pois_mean = 100,
  beta_alpha_1 = 0.01,
  beta_alpha_2 = 0.5,
  seed = NULL
)
}
\arguments{
\item{pois_mean}{Average number of events per game}

\item{beta_alpha_1}{First parameter that affects probability events become goals}

\item{beta_alpha_2}{Second parameter that affects probability events become goals}

\item{seed}{Set seed to enable identical simulation across calls}
}
\description{
Simulates goals as described in Chapter 1
}
\examples{
goals_sim()

}
