\name{PWS-package}
\alias{PWS-package}
\alias{PWS}
\docType{package}
\title{
Personal Weather Station finder
}
\description{
This package implements the capbilities specified in the "PWS" assignment
described in STATS 290.
}
\details{
\tabular{ll}{
Package: \tab PWS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-03-12\cr
License: \tab LGPL-2.1\cr
LazyLoad: \tab yes\cr
}
The user will specify a location and a radius. This package implements functions
to find a list of all personal weather stations within that radius and returns
weather report for each one.
}
\author{
Steve Pittard
Maintainer: Steve Pittard <yourfault@somewhere.net>
}
\references{
STATS290 Website
}
\keyword{ package }
\seealso{
None
}
\examples{
}
