% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_by_sample_size.R
\name{se_by_sample_size}
\alias{se_by_sample_size}
\title{Standard errors and sample size}
\usage{
se_by_sample_size(
  lambda = 2.5,
  ns = c(10, 50, 100, 500, 1000, 5000, 10000),
  n_resample = 10000,
  transform_n = FALSE,
  seed = NULL
)
}
\arguments{
\item{lambda}{Poisson population mean}

\item{ns}{vector of sample sizes}

\item{n_resample}{number of resamples}

\item{transform_n}{(TRUE/FALSE) transform x axis of plot?}

\item{seed}{seed for random number generator}
}
\value{
plot of standard error of estimate versus sample size
}
\description{
Illustration of effect of sample size on standard error
}
\examples{
se_by_sample_size()

}
