% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xG_dist_data.R
\docType{data}
\name{xG_dist_data}
\alias{xG_dist_data}
\title{Simulated data for xG analysis}
\format{
A data frame with 5000 rows and 2 columns
Each row corresponds to one goalscoring opportunity
The columns give distance from gol and a binary indicator of whether a goal was scored or not
}
\usage{
xG_dist_data
}
\description{
distance and goal scored indicator for 5000 simulated goalscooring opportunities
}
\examples{
head(xG_dist_data, 10) # Frst 10 rows of data
mean(xG_dist_data$goal) # Average xG value across all distances
}
\keyword{datasets}
